Eric Sunshine eliminated some more compilation warnings in lighter2 reported by a more recent gcc. Also normalized formatting of size_t per suggestion from Frank.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@28206 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/tools/lighter2/primitive.cpp
@@ -198,9 +198,9 @@ namespace lighter
     for (uint i = fy; i < ly; ++i)
     {
       if (x < scanBuf[i].min)
-        scanBuf[i].min = x;
+        scanBuf[i].min = static_cast<uint>(x);
       if ((x+1) > scanBuf[i].max)
-        scanBuf[i].max = (x+1);
+        scanBuf[i].max = static_cast<uint>(x+1);
 
       x += slopeX;
     }
